make: Fix symbol resolving for binutils-2.21
We cannot override symbols defined as ABS in libl4re-vfs. Since the library and us
use the same definition from l4defs.mk.inc, we can safely remove the override and
use the implicit assumption that the values will match.

I know that this is a hack, maybe we should check for the actual value from
libl4re-vfs in a later "real" fix.